code splitting resolver
