+++ app/cmd/main.fer
@@ -92,3 +92,10 @@ let laterDeclaredValue = "This value is declared later in the code.";
 //array 
 //let numbers = [1, 2, 3, 4, 5]; // Array of integers
 // let mixedArray = [1, "two", 3.0]; // Array
+
+//function
+fn nothing() {
+    // This function does nothing
+}
+
+nothing(); // Call the function
