+++ compiler/cmd/flags.go
@@ -1,44 +1,41 @@
 package main
 
 import (
-	"flag"
 	"os"
 )
 
 func parseArgs() (string, bool, bool, string, string) {
 	var debug bool
 	var outputPath string
-
-	// Define flags
-	flag.BoolVar(&debug, "debug", false, "Enable debug mode")
-	flag.StringVar(&outputPath, "o", "", "Output path for compiled assembly")
-	flag.StringVar(&outputPath, "output", "", "Output path for compiled assembly")
-
-	// Custom usage function
-	flag.Usage = func() {
-		flag.CommandLine.SetOutput(os.Stderr)
-		flag.PrintDefaults()
-	}
-
-	// Parse flags
-	flag.Parse()
-
-	// Get remaining arguments
-	args := flag.Args()
-
 	var filename string
 	var initProject bool
 	var initPath string
 
-	// Handle subcommands and positional arguments
-	if len(args) > 0 {
-		if args[0] == "init" {
-			initProject = true
-			if len(args) > 1 {
-				initPath = args[1]
+	args := os.Args[1:]
+
+	// Parse arguments manually to handle mixed flag and positional argument order
+	for i := 0; i < len(args); i++ {
+		arg := args[i]
+		switch arg {
+		case "-debug":
+			debug = true
+		case "-o", "-output":
+			if i+1 < len(args) {
+				outputPath = args[i+1]
+				i++ // Skip the next argument since we consumed it
 			}
-		} else {
-			filename = args[0]
+		case "init":
+			initProject = true
+			// Check if next argument is a path
+			if i+1 < len(args) && args[i+1][:1] != "-" {
+				initPath = args[i+1]
+				i++
+			}
+		default:
+			// If it's not a flag and we haven't set filename yet, this is the filename
+			if !initProject && filename == "" && arg[:1] != "-" {
+				filename = arg
+			}
 		}
 	}
